refactor: extract component from page
